Renamed package to dds_cli
