+++ dds_cli/__main__.py
@@ -22,13 +22,13 @@ import rich.prompt
 import click_pathlib
 
 # Own modules
-import dds
-from dds import directory
-from dds import timestamp
-from dds import data_putter as dp
-from dds import data_lister as dl
-from dds import data_remover as dr
-from dds import data_getter as dg
+import dds_cli
+from dds_cli import directory
+from dds_cli import timestamp
+from dds_cli import data_putter as dp
+from dds_cli import data_lister as dl
+from dds_cli import data_remover as dr
+from dds_cli import data_getter as dg
 
 ###############################################################################
 # START LOGGING CONFIG ################################# START LOGGING CONFIG #
@@ -118,7 +118,7 @@ def dds_cli(ctx, debug):
 ###############################################################################
 
 
-@cli.command()
+@dds_cli.command()
 @click.option(
     "--config",
     "-c",
@@ -322,7 +322,7 @@ def put(
 ###############################################################################
 
 
-@cli.command()
+@dds_cli.command()
 @click.argument("fold_arg", required=False)  # Needs to be before proj_arg
 @click.argument("proj_arg", required=False)
 @click.option("--project", "-p", required=False, help="Project ID.")
@@ -387,7 +387,7 @@ def ls(dds_info, proj_arg, fold_arg, project, projects, folder, size, username, 
 ###############################################################################
 
 
-@cli.command()
+@dds_cli.command()
 @click.argument("proj_arg", required=False)
 @click.option("--project", required=False, type=str, help="Project ID.")
 @click.option(
@@ -476,7 +476,7 @@ def rm(dds_info, proj_arg, project, username, rm_all, file, folder, config):
 ###############################################################################
 
 
-@cli.command()
+@dds_cli.command()
 @click.option(
     "--config",
     "-c",
