+++ dds_cli/text_handler.py
